gunpai/requirements.txt
attrs==24.2.0
auth==0.5.3
blinker==1.9.0
cattrs==24.1.2
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
coremltools==8.0
cycler==0.12.1
dnspython==2.7.0
eventlet==0.38.0
falcon==4.0.2
filelock==3.16.1
Flask==3.1.0
Flask-Login==0.6.3
fonttools==4.54.1
fsspec==2024.10.0
future==1.0.0
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
iso8601==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
mongoengine==0.29.1
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
onnx==1.17.0
onnx-coreml==1.3
opencv-python==4.10.0.84
packaging==24.1
paho-mqtt==2.1.0
pandas==2.2.3
pillow==10.4.0
protobuf==5.28.3
psutil==6.1.0
py-cpuinfo==9.0.0
pyaml==24.9.0
pymongo==4.10.1
pyparsing==3.2.0
PyPDF2==3.0.1
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
pyzbar==0.1.9
reportlab==4.2.5
requests==2.32.3
scipy==1.14.1
seaborn==0.13.2
serial==0.0.97
six==1.16.0
sympy==1.13.1
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.66.6
typing==3.7.4.3
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.24
ultralytics-thop==2.0.9
urllib3==2.2.3
Werkzeug==3.1.3
